💈👾 -> SSDROP-3: Fixing the 👔Stripe frontend and adding 🍵Mint component
https://github.com/gizmotronn/mint/issues/8

DAO-entry.js is the app.js of the stripe frontend. I've added the app.js of the mint (from gizmotronn/mint (branch: main)) (not the index.js yet) and started going about adding the components into the app (so that it combines into one single react project). Yarn/npm modules/libraries are being imported

Currently this app will not run - we need to update index.js in "root" of this branch on this repo to work for the mint & the dao-entry.js file

the "client" folder from "profile" branch has been sent to the root and the server folder has been sent to a server folder (i.e. taken out of root). Running yarn start in both new locations still works.